moved c3 to parent dir and remove panics
