+++ crates/solidity/outputs/cargo/crate/src/backend/mod.rs
@@ -1,6 +1,8 @@
+mod c3;
+
 pub mod l1_structured_ast;
 pub mod l2_flat_contracts;
 
 pub mod binder;
-pub mod types;
 pub mod passes;
+pub mod types;
