+++ crates/solidity/outputs/cargo/crate/src/backend/passes/mod.rs
@@ -1,5 +1,3 @@
-mod c3;
-
 pub mod p0_build_ast;
 pub mod p1_flatten_contracts;
 pub mod p2_collect_definitions;
